refactor: IFS-2416: move classes to security package
