+++ isy-task/src/main/java/de/bund/bva/isyfact/task/security/AuthenticatorFactory.java
@@ -1,4 +1,4 @@
-package de.bund.bva.isyfact.task.sicherheit;
+package de.bund.bva.isyfact.task.security;
 
 /**
  * Interface über das Instanzen von {@link Authenticator} erzeugt werden.
