+++ isy-task/src/main/java/de/bund/bva/isyfact/task/security/impl/NoOpAuthenticator.java
@@ -1,6 +1,6 @@
-package de.bund.bva.isyfact.task.sicherheit.impl;
+package de.bund.bva.isyfact.task.security.impl;
 
-import de.bund.bva.isyfact.task.sicherheit.Authenticator;
+import de.bund.bva.isyfact.task.security.Authenticator;
 
 /**
  * Implementierung von {@link Authenticator}, die eine "leere" Authentifizierung durchführt.
